Rework values.
Keep a list of 'values handlers' that, when needed, update the values
object somehow. They're all called in sequence when the values property
is accessed; the list of handlers can be manipulated during eg setup()
or pre_run(). This should make it easier to support other things like
OS X plists or the Windows registry.