Make sure the CLI handler doesn't set a default value.
Previously, if a parameter with a default value was passed to the CLI handler,
it would convert that parameter to an Option instance with a default value. This
makes it difficult to decide whether or not to update the parameter's value attribute
after handle(). Instead, pass a bogus value as default and, if we see that when
handling parameters, stick with the parameter's default.